LPS-28938 Rename package and source formatting
