+++ portal-impl/test/integration/com/liferay/counter/service/CounterLocalServiceTest.java
@@ -16,10 +16,10 @@ package com.liferay.counter.service;
 
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.service.PortalServiceUtil;
-import com.liferay.portal.service.ServiceTestUtil;
 import com.liferay.portal.test.EnvironmentExecutionTestListener;
 import com.liferay.portal.test.ExecutionTestListeners;
 import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
+import com.liferay.portal.test.service.ServiceTestUtil;
 
 import org.junit.Assert;
 import org.junit.Test;
