+++ portal-impl/test/integration/com/liferay/portal/service/OrganizationLocalServiceTest.java
@@ -23,6 +23,7 @@ import com.liferay.portal.test.EnvironmentExecutionTestListener;
 import com.liferay.portal.test.ExecutionTestListeners;
 import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
 import com.liferay.portal.test.TransactionalExecutionTestListener;
+import com.liferay.portal.test.service.ServiceTestUtil;
 
 import java.util.List;
