+++ portal-impl/test/integration/com/liferay/portal/service/ResourceLocalServiceTest.java
@@ -22,7 +22,8 @@ import com.liferay.portal.security.permission.DoAsUserThread;
 import com.liferay.portal.test.EnvironmentExecutionTestListener;
 import com.liferay.portal.test.ExecutionTestListeners;
 import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
-import com.liferay.portal.util.TestPropsValues;
+import com.liferay.portal.test.service.ServiceTestUtil;
+import com.liferay.portal.test.util.TestPropsValues;
 
 import org.junit.Assert;
 import org.junit.Before;
