+++ portal-impl/test/integration/com/liferay/portlet/bookmarks/service/BaseBookmarksServiceTestCase.java
@@ -15,7 +15,7 @@
 package com.liferay.portlet.bookmarks.service;
 
 import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.util.TestPropsValues;
+import com.liferay.portal.test.util.TestPropsValues;
 import com.liferay.portlet.bookmarks.model.BookmarksEntry;
 import com.liferay.portlet.bookmarks.model.BookmarksFolder;
 import com.liferay.portlet.bookmarks.model.BookmarksFolderConstants;
