+++ portal-impl/test/integration/com/liferay/portlet/documentlibrary/store/AdvancedFileSystemStoreTest.java
@@ -15,10 +15,10 @@
 package com.liferay.portlet.documentlibrary.store;
 
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.service.ServiceTestUtil;
 import com.liferay.portal.test.EnvironmentExecutionTestListener;
 import com.liferay.portal.test.ExecutionTestListeners;
 import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
+import com.liferay.portal.test.service.ServiceTestUtil;
 import com.liferay.portlet.documentlibrary.NoSuchDirectoryException;
 
 import org.junit.Assert;
