+++ portal-impl/test/integration/com/liferay/portlet/documentlibrary/store/DBStoreTest.java
@@ -17,11 +17,11 @@ package com.liferay.portlet.documentlibrary.store;
 import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayInputStream;
 import com.liferay.portal.kernel.transaction.Transactional;
 import com.liferay.portal.kernel.util.SetUtil;
-import com.liferay.portal.service.ServiceTestUtil;
 import com.liferay.portal.test.EnvironmentExecutionTestListener;
 import com.liferay.portal.test.ExecutionTestListeners;
 import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
 import com.liferay.portal.test.TransactionalExecutionTestListener;
+import com.liferay.portal.test.service.ServiceTestUtil;
 
 import java.io.BufferedInputStream;
 import java.io.ByteArrayInputStream;
